+++ examples/include.js
@@ -1,5 +1,9 @@
-import jazz from "../lib/jazz";
+import jazz from "jazz";
 import fs from "fs";
+import path from "path";
+import { fileURLToPath } from "url";
+
+const __dirname = path.dirname(fileURLToPath(import.meta.url));
 
 const data = fs.readFileSync(__dirname + "/include_main.jazz", "utf8");
 const template = jazz.compile(data);